+++ packages/file-upload/README.md
@@ -1,0 +1,34 @@
+---
+title: FileUpload
+description:
+group: Forms
+---
+
+
+## Accepted files
+
+- `audio/*`
+- `audio/mpeg`
+- `audio/wav`
+- `image/*`
+- `image/gif`
+- `image/heic`
+- `image/jpeg`
+- `image/png`
+- `image/svg+xml`
+- `image/tiff`
+- `image/webp`
+- `text/*`
+- `text/csv`
+- `text/plain`
+- `text/rtf`
+- `video/*`
+- `video/mp4`
+- `video/mpeg`
+- `application/msword`
+- `application/pdf`
+- `application/rtf`
+- `application/vnd.ms-excel`
+- `application/vnd.openxmlformats-officedocument.spreadsheetml.sheet` (.xls)
+- `application/vnd.openxmlformats-officedocument.wordprocessingml.document` (.doc)
+- `application/zip`
